+++ lib/ruby_lokalise_api/base_client.rb
@@ -2,11 +2,11 @@
 
 module RubyLokaliseApi
   # This class contains the base client. Inherited by Client (regular API client)
-  # and OAuth2Client (used for OAuth-2 based authentication)
+  # and OAuth2Client (used for OAuth2-based authentication)
   class BaseClient
     include RubyLokaliseApi::Rest
 
-    attr_reader :token, :token_header
+    attr_reader :token, :token_header, :api_host
     attr_accessor :timeout, :open_timeout
 
     def initialize(token, params = {})
@@ -14,6 +14,7 @@ module RubyLokaliseApi
       @timeout = params.fetch(:timeout, nil)
       @open_timeout = params.fetch(:open_timeout, nil)
       @token_header = ''
+      @api_host = params.fetch(:api_host, nil)
     end
   end
 end